requirements.txt
alembic==1.4.2
bcrypt==3.1.7
cffi==1.14.0
click==7.1.2
dnspython==1.16.0
dominate==2.5.1
email-validator==1.1.1
environs==7.4.0
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.2
MarkupSafe==1.1.1
marshmallow==3.6.0
psycopg2-binary==2.8.5
pycparser==2.20
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
python-slugify==4.0.0
six==1.15.0
SQLAlchemy==1.3.17
text-unidecode==1.3
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.3.1
